move setup to  pyproject.toml
