+++ pyproject.toml
@@ -1,6 +1,76 @@
 [build-system]
-requires = ["setuptools", "wheel", "nltk==3.9.1", "textblob==0.19.0"]
+requires = ["setuptools>=61", "wheel", "nltk==3.9.1", "textblob==0.19.0"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+include = ["duckbot", "scripts"]
+
+[project]
+name = "duckbot"
+version = "1.0"
+description = ""
+readme = "README.md"
+requires-python = ">=3.10"
+urls = { Homepage = "https://github.com/duck-dynasty/duckbot" }
+
+dependencies = [
+    "discord.py[voice]==2.5.0",
+    "beautifulsoup4==4.13.4",
+    "requests==2.32.3",
+    "pytz==2025.2",
+    "timezonefinder==6.5.9",
+    "holidays==0.73",
+    "pyowm==3.3.0",
+    "psycopg2==2.9.10",
+    "SQLAlchemy==2.0.41",
+    "d20==1.1.2",
+    "nltk==3.9.1",
+    "textblob==0.19.0",
+    "pyfiglet==1.0.2",
+    "matplotlib==3.10.3",
+    "PyGithub==2.1.1",
+    "wolframalpha==5.1.3",
+    "yfinance==0.2.61",
+    "mip==1.15.0",
+    "anthropic==0.52.1"
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest==8.3.5",
+    "pytest-asyncio==0.26.0",
+    "pytest-xdist[psutil]==3.7.0",
+    "flake8==7.2.0",
+    "black==25.1.0",
+    "flake8-black==0.3.6",
+    "isort==6.0.1",
+    "flake8-isort==6.1.2",
+    "pep8-naming==0.15.1",
+    "mdformat==0.7.22",
+    "mdformat-gfm==0.4.1",
+    "mdformat-black==0.1.1",
+    "responses==0.25.7",
+    "pytest-blockage==0.2.4",
+    "pytest-sugar==1.0.0",
+    "pytest-icdiff==0.9",
+    "pytest-cov==6.1.1"
+]
+
+cdk = [
+    "aws-cdk.core==1.204.0",
+    "aws-cdk.aws-ec2==1.204.0",
+    "aws-cdk.aws-ecs==1.204.0",
+    "aws-cdk.aws-autoscaling==1.204.0",
+    "aws-cdk.aws-efs==1.204.0",
+    "aws-cdk.aws-iam==1.204.0",
+    "aws-cdk.aws-logs==1.204.0",
+    "aws-cdk.aws-ssm==1.204.0",
+    "boto3==1.38.27"
+]
+
+[project.scripts]
+format = "scripts.cli:format"
+setup_nltk = "scripts.cli:setup_nltk"
 
 [tool.black]
 line-length = 200
@@ -16,3 +86,8 @@ addopts = """\
     -n auto --dist loadfile \
     --blockage"""
 asyncio_mode = "auto"
+
+[tool.flake8]
+max-line-length = 200
+max-complexity = 10
+per-file-ignores = ["__init__.py:F401"]
