--- setup.cfg
@@ -1,6 +1,0 @@
-[flake8]
-# ignore unused imports in __init__.py files
-per-file-ignores = __init__.py:F401
-
-max-complexity = 10
-max-line-length = 200
